+++ Sources/ArgumentParser/Parsable Properties/Argument.swift
@@ -290,7 +290,7 @@ public struct ArgumentArrayParsingStrategy: Hashable {
 }
 
 // MARK: - @Argument T: ExpressibleByArgument Initializers
-extension Argument {
+extension Argument where Value: ExpressibleByArgument {
   /// Creates a property with a default value provided by standard Swift default
   /// value syntax.
   ///
@@ -309,7 +309,7 @@ extension Argument {
     wrappedValue: Value,
     help: ArgumentHelp? = nil,
     completion: CompletionKind? = nil
-  ) where Value: ExpressibleByArgument {
+  ) {
     self.init(_parsedValue: .init { key in
       let arg = ArgumentDefinition(
         container: Bare<Value>.self,
@@ -338,7 +338,7 @@ extension Argument {
   public init(
     help: ArgumentHelp? = nil,
     completion: CompletionKind? = nil
-  ) where Value: ExpressibleByArgument {
+  ) {
     self.init(_parsedValue: .init { key in
       let arg = ArgumentDefinition(
         container: Bare<Value>.self,
@@ -459,8 +459,9 @@ extension Argument {
   @available(*, deprecated, message: """
     Optional @Arguments with default values should be declared as non-Optional.
     """)
+  @_disfavoredOverload
   public init<T>(
-    wrappedValue: Optional<T>,
+    wrappedValue _wrappedValue: Optional<T>,
     help: ArgumentHelp? = nil,
     completion: CompletionKind? = nil
   ) where T: ExpressibleByArgument, Value == Optional<T> {
@@ -471,7 +472,7 @@ extension Argument {
         kind: .positional,
         help: help,
         parsingStrategy: .default,
-        initial: wrappedValue,
+        initial: _wrappedValue,
         completion: completion)
 
       return ArgumentSet(arg)
@@ -540,8 +541,9 @@ extension Argument {
   @available(*, deprecated, message: """
     Optional @Arguments with default values should be declared as non-Optional.
     """)
+  @_disfavoredOverload
   public init<T>(
-    wrappedValue: Optional<T>,
+    wrappedValue _wrappedValue: Optional<T>,
     help: ArgumentHelp? = nil,
     completion: CompletionKind? = nil,
     transform: @escaping (String) throws -> T
@@ -554,7 +556,7 @@ extension Argument {
         help: help,
         parsingStrategy: .default,
         transform: transform,
-        initial: wrappedValue,
+        initial: _wrappedValue,
         completion: completion)
 
       return ArgumentSet(arg)