+++ Sources/ArgumentParser/Parsable Types/ParsableArguments.swift
@@ -9,19 +9,6 @@
 //
 //===----------------------------------------------------------------------===//
 
-#if canImport(Glibc)
-import Glibc
-let _exit: (Int32) -> Never = Glibc.exit
-#elseif canImport(Darwin)
-import Darwin
-let _exit: (Int32) -> Never = Darwin.exit
-#elseif canImport(CRT)
-import CRT
-let _exit: (Int32) -> Never = ucrt._exit
-#elseif canImport(WASILibc)
-import WASILibc
-#endif
-
 /// A type that can be parsed from a program's command-line arguments.
 ///
 /// When you implement a `ParsableArguments` type, all properties must be declared with
@@ -62,14 +49,6 @@ struct _WrappedParsableCommand<P: ParsableArguments>: ParsableCommand {
 
   @OptionGroup var options: P
 }
-
-struct StandardError: TextOutputStream {
-  mutating func write(_ string: String) {
-    for byte in string.utf8 { putc(numericCast(byte), stderr) }
-  }
-}
-
-var standardError = StandardError()
 
 extension ParsableArguments {
   public mutating func validate() throws {}
@@ -209,7 +188,7 @@ extension ParsableArguments {
     withError error: Error? = nil
   ) -> Never {
     guard let error = error else {
-      _exit(ExitCode.success.rawValue)
+      Platform.exit(ExitCode.success.rawValue)
     }
     
     let messageInfo = MessageInfo(error: error, type: self)
@@ -218,10 +197,10 @@ extension ParsableArguments {
       if messageInfo.shouldExitCleanly {
         print(fullText)
       } else {
-        print(fullText, to: &standardError)
+        print(fullText, to: &Platform.standardError)
       }
     }
-    _exit(messageInfo.exitCode.rawValue)
+    Platform.exit(messageInfo.exitCode.rawValue)
   }
   
   /// Parses a new instance of this type from command-line arguments or exits
@@ -270,7 +249,7 @@ extension ArgumentSetProvider {
 }
 
 extension ArgumentSet {
-  init(_ type: ParsableArguments.Type, visibility: ArgumentVisibility, parent: InputKey.Parent) {
+  init(_ type: ParsableArguments.Type, visibility: ArgumentVisibility, parent: InputKey?) {
     #if DEBUG
     do {
       try type._validate(parent: parent)