+++ Sources/ArgumentParser/Parsing/ArgumentDefinition.swift
@@ -218,7 +218,7 @@ extension ArgumentDefinition {
   ///
   /// This initializer is used for any property defined on a `ParsableArguments`
   /// type that isn't decorated with one of ArgumentParser's property wrappers.
-  init(unparsedKey: String, default defaultValue: Any?, parent: InputKey.Parent) {
+  init(unparsedKey: String, default defaultValue: Any?, parent: InputKey?) {
     self.init(
       container: Bare<Any>.self,
       key: InputKey(name: unparsedKey, parent: parent),