+++ Sources/ArgumentParser/Usage/HelpGenerator.swift
@@ -12,7 +12,7 @@
 internal struct HelpGenerator {
   static var helpIndent = 2
   static var labelColumnWidth = 26
-  static var systemScreenWidth: Int { _terminalSize().width }
+  static var systemScreenWidth: Int { Platform.terminalWidth }
 
   struct Section {
     struct Element: Hashable {
@@ -97,7 +97,7 @@ internal struct HelpGenerator {
       fatalError()
     }
     
-    let currentArgSet = ArgumentSet(currentCommand, visibility: visibility, parent: .root)
+    let currentArgSet = ArgumentSet(currentCommand, visibility: visibility, parent: nil)
     self.commandStack = commandStack
 
     // Build the tool name and subcommand name from the command configuration
@@ -292,7 +292,7 @@ fileprivate extension NameSpecification {
   /// step, the name are returned in descending order.
   func generateHelpNames(visibility: ArgumentVisibility) -> [Name] {
     self
-      .makeNames(InputKey(name: "help", parent: .root))
+      .makeNames(InputKey(name: "help", parent: nil))
       .compactMap { name in
         guard visibility.base != .default else { return name }
         switch name {
@@ -333,7 +333,7 @@ internal extension BidirectionalCollection where Element == ParsableCommand.Type
         options: [.isOptional],
         help: "Show the version.",
         defaultValue: nil,
-        key: InputKey(name: "", parent: .root),
+        key: InputKey(name: "", parent: nil),
         isComposite: false),
       completion: .default,
       update: .nullary({ _, _, _ in })
@@ -350,7 +350,7 @@ internal extension BidirectionalCollection where Element == ParsableCommand.Type
         options: [.isOptional],
         help: "Show help information.",
         defaultValue: nil,
-        key: InputKey(name: "", parent: .root),
+        key: InputKey(name: "", parent: nil),
         isComposite: false),
       completion: .default,
       update: .nullary({ _, _, _ in })
@@ -365,7 +365,7 @@ internal extension BidirectionalCollection where Element == ParsableCommand.Type
         options: [.isOptional],
         help: ArgumentHelp("Dump help information as JSON."),
         defaultValue: nil,
-        key: InputKey(name: "", parent: .root),
+        key: InputKey(name: "", parent: nil),
         isComposite: false),
       completion: .default,
       update: .nullary({ _, _, _ in })
@@ -375,7 +375,7 @@ internal extension BidirectionalCollection where Element == ParsableCommand.Type
   /// Returns the ArgumentSet for the last command in this stack, including
   /// help and version flags, when appropriate.
   func argumentsForHelp(visibility: ArgumentVisibility) -> ArgumentSet {
-    guard var arguments = self.last.map({ ArgumentSet($0, visibility: visibility, parent: .root) })
+    guard var arguments = self.last.map({ ArgumentSet($0, visibility: visibility, parent: nil) })
       else { return ArgumentSet() }
     self.versionArgumentDefinition().map { arguments.append($0) }
     self.helpArgumentDefinition().map { arguments.append($0) }
@@ -385,44 +385,4 @@ internal extension BidirectionalCollection where Element == ParsableCommand.Type
     
     return arguments
   }
-}
-
-#if canImport(Glibc)
-import Glibc
-func ioctl(_ a: Int32, _ b: Int32, _ p: UnsafeMutableRawPointer) -> Int32 {
-  ioctl(CInt(a), UInt(b), p)
-}
-#elseif canImport(Darwin)
-import Darwin
-#elseif canImport(CRT)
-import CRT
-import WinSDK
-#endif
-
-func _terminalSize() -> (width: Int, height: Int) {
-#if os(WASI)
-  // WASI doesn't yet support terminal size
-  return (80, 25)
-#elseif os(Windows)
-  var csbi: CONSOLE_SCREEN_BUFFER_INFO = CONSOLE_SCREEN_BUFFER_INFO()
-  guard GetConsoleScreenBufferInfo(GetStdHandle(STD_OUTPUT_HANDLE), &csbi) else {
-    return (80, 25)
-  }
-  return (width: Int(csbi.srWindow.Right - csbi.srWindow.Left) + 1,
-          height: Int(csbi.srWindow.Bottom - csbi.srWindow.Top) + 1)
-#else
-  var w = winsize()
-#if os(OpenBSD)
-  // TIOCGWINSZ is a complex macro, so we need the flattened value.
-  let tiocgwinsz = Int32(0x40087468)
-  let err = ioctl(STDOUT_FILENO, tiocgwinsz, &w)
-#else
-  let err = ioctl(STDOUT_FILENO, TIOCGWINSZ, &w)
-#endif
-  let width = Int(w.ws_col)
-  let height = Int(w.ws_row)
-  guard err == 0 else { return (80, 25) }
-  return (width: width > 0 ? width : 80,
-          height: height > 0 ? height : 25)
-#endif
 }