+++ Sources/ArgumentParser/Usage/UsageGenerator.swift
@@ -9,7 +9,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-@_implementationOnly import Foundation
+@_implementationOnly import protocol Foundation.LocalizedError
 
 struct UsageGenerator {
   var toolName: String
@@ -22,7 +22,7 @@ extension UsageGenerator {
     self.init(toolName: toolName, definition: definition)
   }
   
-  init(toolName: String, parsable: ParsableArguments, visibility: ArgumentVisibility, parent: InputKey.Parent) {
+  init(toolName: String, parsable: ParsableArguments, visibility: ArgumentVisibility, parent: InputKey?) {
     self.init(
       toolName: toolName,
       definition: ArgumentSet(type(of: parsable), visibility: visibility, parent: parent))
@@ -115,6 +115,9 @@ extension ArgumentDefinition {
     }
     if help.options.contains(.isOptional) {
       synopsis = "[\(synopsis)]"
+    }
+    if parsingStrategy == .postTerminator {
+      synopsis = "-- \(synopsis)"
     }
     return synopsis
   }