+++ Tests/ArgumentParserUnitTests/HelpGenerationTests+AtArgument.swift
@@ -221,6 +221,7 @@ extension HelpGenerationTests {
       var arg0: A? = nil
     }
 
+    @available(*, deprecated, message: "Included for test coverage")
     struct OptionalDefault: ParsableCommand {
       @Argument(help: "example")
       var arg0: A? = A()