+++ Tests/ArgumentParserUnitTests/HelpGenerationTests.swift
@@ -135,15 +135,15 @@ extension HelpGenerationTests {
 
   enum OptionFlags: String, EnumerableFlag { case optional, required }
   enum Degree {
-    case bachelor, master, doctor
+    case bachelor, graduate, doctorate
     static func degreeTransform(_ string: String) throws -> Degree {
       switch string {
       case "bachelor":
         return .bachelor
-      case "master":
-        return .master
-      case "doctor":
-        return .doctor
+      case "graduate":
+        return .graduate
+      case "doctorate":
+        return .doctorate
       default:
         throw ValidationError("Not a valid string for 'Degree'")
       }
@@ -365,8 +365,8 @@ extension HelpGenerationTests {
   func testOverviewButNoAbstractSpacing() {
     let renderedHelp = HelpGenerator(J.self, visibility: .default)
       .rendered()
-    AssertEqualStringsIgnoringTrailingWhitespace(renderedHelp, """
-    OVERVIEW:
+    AssertEqualStrings(actual: renderedHelp, expected: """
+    OVERVIEW: \n\
     test
 
     USAGE: j
@@ -640,7 +640,7 @@ extension HelpGenerationTests {
   }
 
   func testAllValues() {
-    let opts = ArgumentSet(AllValues.self, visibility: .private, parent: .root)
+    let opts = ArgumentSet(AllValues.self, visibility: .private, parent: nil)
     XCTAssertEqual(AllValues.Manual.allValueStrings, opts[0].help.allValues)
     XCTAssertEqual(AllValues.Manual.allValueStrings, opts[1].help.allValues)