+++ Tests/ArgumentParserUnitTests/ParsableArgumentsValidationTests.swift
@@ -84,7 +84,7 @@ final class ParsableArgumentsValidationTests: XCTestCase {
   }
 
   func testCodingKeyValidation() throws {
-    let parent = InputKey.Parent.key(InputKey(name: "parentKey", parent: .root))
+    let parent = InputKey(name: "parentKey", parent: nil)
     XCTAssertNil(ParsableArgumentsCodingKeyValidator.validate(A.self, parent: parent))
     XCTAssertNil(ParsableArgumentsCodingKeyValidator.validate(B.self, parent: parent))
 
@@ -133,7 +133,7 @@ final class ParsableArgumentsValidationTests: XCTestCase {
   }
 
   func testCustomDecoderValidation() throws {
-    let parent = InputKey.Parent(InputKey(name: "foo", parent: .root))
+    let parent = InputKey(name: "foo", parent: nil)
         if let error = ParsableArgumentsCodingKeyValidator.validate(TypeWithInvalidDecoder.self, parent: parent)
       as? ParsableArgumentsCodingKeyValidator.InvalidDecoderError
     {
@@ -214,7 +214,7 @@ final class ParsableArgumentsValidationTests: XCTestCase {
   }
 
   func testPositionalArgumentsValidation() throws {
-    let parent = InputKey.Parent(InputKey(name: "foo", parent: .root))
+    let parent = InputKey(name: "foo", parent: nil)
     XCTAssertNil(PositionalArgumentsValidator.validate(A.self, parent: parent))
     XCTAssertNil(PositionalArgumentsValidator.validate(F.self, parent: parent))
     XCTAssertNil(PositionalArgumentsValidator.validate(H.self, parent: parent))
@@ -249,7 +249,7 @@ final class ParsableArgumentsValidationTests: XCTestCase {
   }
 
   func testUniqueNamesValidation_NoViolation() throws {
-    let parent = InputKey.Parent(InputKey(name: "foo", parent: .root))
+    let parent = InputKey(name: "foo", parent: nil)
     XCTAssertNil(ParsableArgumentsUniqueNamesValidator.validate(DifferentNames.self, parent: parent))
   }
 
@@ -264,7 +264,7 @@ final class ParsableArgumentsValidationTests: XCTestCase {
   }
 
   func testUniqueNamesValidation_TwoOfSameName() throws {
-    if let error = ParsableArgumentsUniqueNamesValidator.validate(TwoOfTheSameName.self, parent: .root)
+    if let error = ParsableArgumentsUniqueNamesValidator.validate(TwoOfTheSameName.self, parent: nil)
       as? ParsableArgumentsUniqueNamesValidator.Error
     {
       XCTAssertEqual(error.description, "Multiple (2) `Option` or `Flag` arguments are named \"--foo\".")
@@ -292,7 +292,7 @@ final class ParsableArgumentsValidationTests: XCTestCase {
   }
 
   func testUniqueNamesValidation_TwoDuplications() throws {
-    let parent = InputKey.Parent(InputKey(name: "option", parent: .root))
+    let parent = InputKey(name: "option", parent: nil)
     if let error = ParsableArgumentsUniqueNamesValidator.validate(MultipleUniquenessViolations.self, parent: parent)
       as? ParsableArgumentsUniqueNamesValidator.Error
     {
@@ -329,7 +329,7 @@ final class ParsableArgumentsValidationTests: XCTestCase {
   }
 
   func testUniqueNamesValidation_ArgumentHasMultipleNames() throws {
-    if let error = ParsableArgumentsUniqueNamesValidator.validate(MultipleNamesPerArgument.self, parent: .root)
+    if let error = ParsableArgumentsUniqueNamesValidator.validate(MultipleNamesPerArgument.self, parent: nil)
       as? ParsableArgumentsUniqueNamesValidator.Error
     {
       XCTAssertEqual(error.description, "Multiple (2) `Option` or `Flag` arguments are named \"-v\".")
@@ -360,7 +360,7 @@ final class ParsableArgumentsValidationTests: XCTestCase {
   }
 
   func testUniqueNamesValidation_MoreThanTwoDuplications() throws {
-    if let error = ParsableArgumentsUniqueNamesValidator.validate(FourDuplicateNames.self, parent: .root)
+    if let error = ParsableArgumentsUniqueNamesValidator.validate(FourDuplicateNames.self, parent: nil)
       as? ParsableArgumentsUniqueNamesValidator.Error
     {
       XCTAssertEqual(error.description, "Multiple (4) `Option` or `Flag` arguments are named \"--foo\".")
@@ -404,7 +404,7 @@ final class ParsableArgumentsValidationTests: XCTestCase {
   }
 
   func testUniqueNamesValidation_DuplicatedFlagFirstLetters_ShortNames() throws {
-    if let error = ParsableArgumentsUniqueNamesValidator.validate(DuplicatedFirstLettersShortNames.self, parent: .root)
+    if let error = ParsableArgumentsUniqueNamesValidator.validate(DuplicatedFirstLettersShortNames.self, parent: nil)
       as? ParsableArgumentsUniqueNamesValidator.Error
     {
       XCTAssertEqual(error.description, "Multiple (3) `Option` or `Flag` arguments are named \"-f\".")
@@ -414,7 +414,7 @@ final class ParsableArgumentsValidationTests: XCTestCase {
   }
 
   func testUniqueNamesValidation_DuplicatedFlagFirstLetters_LongNames() throws {
-    XCTAssertNil(ParsableArgumentsUniqueNamesValidator.validate(DuplicatedFirstLettersLongNames.self, parent: .root))
+    XCTAssertNil(ParsableArgumentsUniqueNamesValidator.validate(DuplicatedFirstLettersLongNames.self, parent: nil))
   }
 
   fileprivate struct HasOneNonsenseFlag: ParsableCommand {
@@ -447,7 +447,7 @@ final class ParsableArgumentsValidationTests: XCTestCase {
   }
 
   func testNonsenseFlagsValidation_OneFlag() throws {
-    if let error = NonsenseFlagsValidator.validate(HasOneNonsenseFlag.self, parent: .root)
+    if let error = NonsenseFlagsValidator.validate(HasOneNonsenseFlag.self, parent: nil)
       as? NonsenseFlagsValidator.Error
     {
       XCTAssertEqual(
@@ -485,7 +485,7 @@ final class ParsableArgumentsValidationTests: XCTestCase {
   }
 
   func testNonsenseFlagsValidation_MultipleFlags() throws {
-    if let error = NonsenseFlagsValidator.validate(MultipleNonsenseFlags.self, parent: .root)
+    if let error = NonsenseFlagsValidator.validate(MultipleNonsenseFlags.self, parent: nil)
         as? NonsenseFlagsValidator.Error
     {
       XCTAssertEqual(